renamed 'eslint.config.mjs' to 'eslint-config.mjs'
